+++ Sources/ZcashLightClientKit/ClosureSynchronizer.swift
@@ -160,6 +160,8 @@ public protocol ClosureSynchronizer {
 
     func estimateBirthdayHeight(for date: Date, completion: @escaping (BlockHeight) -> Void)
 
+    func httpRequestOverTor(for request: URLRequest, retryLimit: UInt8, completion: @escaping (Result<(data: Data, response: HTTPURLResponse), Error>) -> Void)
+    
     /*
      It can be missleading that these two methods are returning Publisher even this protocol is closure based. Reason is that Synchronizer doesn't
      provide different implementations for these two methods. So Combine it is even here.